+++ packages/zkdb/src/sdk/interfaces/context.ts
@@ -1,4 +1,5 @@
 /* eslint-disable no-unused-vars */
+
 import { MinaBlockchain } from "./blockchain.js";
 import { ZKDatabase } from "./database.js";
 import { GlobalContext } from "./global-context.js";
