+++ packages/zkdb/src/sdk/interfaces/database.ts
@@ -1,11 +1,10 @@
 /* eslint-disable no-unused-vars */
-import { PublicKey } from 'o1js';
 import { ZKCollection } from './collection.js';
-import { GroupQuery } from './group.js';
-import { DatabaseSettings } from '../../../types/database.js';
-import { GroupDescription } from '../../../sdk/types/group.js';
-import { DocumentEncoded, SchemaDefinition } from '../../../sdk/schema.js';
-import { Permissions } from '../../../types/permission.js';
+import { DatabaseSettings } from '../../types/database.js';
+import { GroupDescription } from '../../types/group.js';
+import { SchemaDefinition } from '../schema.js';
+import { Permissions } from '../../types/permission.js';
+import { ZKGroup } from './group.js';
 
 export interface ZKDatabase {
   collection(name: string): ZKCollection;
@@ -22,7 +21,7 @@ export interface ZKDatabase {
   ): Promise<void>;
   newGroup(groupName: string, description: string): Promise<void>;
 
-  group(groupName: string): GroupQuery;
+  group(groupName: string): ZKGroup;
 
   groups(): Promise<GroupDescription[]>;
   getSettings(): Promise<DatabaseSettings>;
