+++ packages/zkdb/src/sdk/interfaces/document.ts
@@ -1,14 +1,8 @@
 /* eslint-disable no-unused-vars */
-import { MerkleWitness } from '../../types/merkle-tree.js';
-import { Ownable } from '../query/interfaces/ownable.js';
-import { DocumentEncoded } from '../schema.js';
+
+import { MerkleWitness } from "../../types/merkle-tree.js";
+import { Ownable } from "./ownable.js";
 
 export interface ZKDocument extends Ownable {
-  getWitness(): Promise<MerkleWitness>;
-  toSchema<
-    T extends {
-      new (..._args: any): InstanceType<T>;
-      deserialize: (_doc: DocumentEncoded) => any;
-    },
-  >(type: T): InstanceType<T>;
+  getWitness(): Promise<MerkleWitness>
 }
