--- packages/zkdb/src/sdk/query/interfaces/query.ts
@@ -1,3 +1,0 @@
-export interface Query<T> {
-  get(): T
-}
