--- packages/zkdb/src/sdk/types/group.ts
@@ -1,6 +1,0 @@
-export type GroupDescription = {
-  name: string,
-  description: string
-  createdBy: string,
-  createdAt: Date
-}
